backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.15.0
paramiko>=3.4.0
